refactor: related-spots API를 별도 컨트롤러/서비스로 분리
